+++ sdl_example/PositionTrackingProcessor.h
@@ -34,7 +34,7 @@ class LFPONLINEAPI PositionTrackingProcessor : public LFPProcessor
 #endif
 
     void detect_positions();
-    SpatialInfo detectPosition(cv::Mat& frame, int timestamp);
+    SpatialInfo detectPosition(cv::Mat& frame, unsigned long long timestamp);
 
     public:
         PositionTrackingProcessor(LFPBuffer *buf);
